Split `ConnectorDir` enum from `Common.h`
