+++ OPHD/Common.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "EnumConnectorDir.h"
+
 #include <NAS2D/Math/Rectangle.h>
 #include <NAS2D/Renderer/Color.h>
 
@@ -83,20 +85,6 @@ enum class IdleReason
 	MineExhausted,
 	MineInactive,
 	InsufficientLuxuryProduct
-};
-
-/**
- * Connector Direction.
- *
- * \note	CONNECTOR_INTERSECTION is intentionally set to '1' to prevent
- *			breaking changes with save files.
- */
-enum ConnectorDir
-{
-	CONNECTOR_INTERSECTION = 1,
-	CONNECTOR_RIGHT,
-	CONNECTOR_LEFT,
-	CONNECTOR_VERTICAL // Functions as an intersection
 };
