+++ OPHD/ophd.vcxproj
@@ -266,6 +266,7 @@
     <ClInclude Include="Constants\Strings.h" />
     <ClInclude Include="Constants\UiConstants.h" />
     <ClInclude Include="DirectionOffset.h" />
+    <ClInclude Include="EnumConnectorDir.h" />
     <ClInclude Include="EnumDirection.h" />
     <ClInclude Include="GraphWalker.h" />
     <ClInclude Include="IOHelper.h" />
