+++ OPHD/ophd.vcxproj.filters
@@ -323,6 +323,9 @@
     <ClInclude Include="DirectionOffset.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="EnumConnectorDir.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="EnumDirection.h">
       <Filter>Header Files</Filter>
     </ClInclude>
